5.1.3: Added the board controller
